add `--yes` flag to delete command

+++ cmd/ksec/delete.go
@@ -14,6 +14,16 @@ var deleteCmd = &cobra.Command{
 }
 
 func deleteCommand(cmd *cobra.Command, args []string) error {
+	skipconfirm, err := cmd.Flags().GetBool("yes")
+	if err != nil {
+		return err
+	}
+
+	if !skipconfirm && !askConfirmation("Are you sure? This action cannot be reversed.") {
+		fmt.Println("canceled")
+		return nil
+	}
+
 	for _, name := range args {
 		if err := secretsClient.Delete(name); err != nil {
 			return err
